Synchronous and asynchronous
Synchronous is added 'Sync' in FileSystems's method.
Asynchronous is "not" added 'Sync' in FileSystmes's method.
Synchronous excute code in order. But Asynchronous excute method
at background, and excute it later calling callback function.
Because, nodejs is single thread. If we use sychronous, bottleneck
occur. So, in this case that method take a long time, using asynchronous
is better.